refactor: Move internal packages
