+++ pkg/cloudencrypt/native_test.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/stretchr/testify/assert"
 
-	"github.com/keboola/go-cloud-encrypt/pkg/cloudencrypt/internal/random"
+	"github.com/keboola/go-cloud-encrypt/internal/random"
 )
 
 func TestNativeEncryptor(t *testing.T) {
